<FIX> Fix a problem due to unconfirmed children
A node sometimes have unconfirmed children. Ideally, receiving the local sensor packet from an unconfirmed child in the first round of request will add it to the list.

However, since the children are unconfirmed, the backoff range is too short (1s) for children to successfully deliver their local sensor in the first round, resulting in a situation where the 1s backoff is used forever and the packet delivery rate is low.

The problem can be fixed by looking at the data collected. Since every request will obtain up to 1 packet from each child, if we find a discrepancy between the number of children and packets received, we can increment the number of children, but do not update the childrenlist yet since the source of the packet might be from other descendants. The increased number of children will likely allow the packets to deliver successfully in the first round of request during the next duty cycle.